# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wa-rev",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/contests/": {
      "get": {
        "tags": [
          "contests"
        ],
        "summary": "Read Contests",
        "description": "Retrieve Contests.",
        "operationId": "contests-read_Contests",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "contests"
        ],
        "summary": "Create Contest",
        "description": "Create new Contest.",
        "operationId": "contests-create_contest",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contests/{id}/problems": {
      "get": {
        "tags": [
          "contests"
        ],
        "summary": "Read Contest Problems",
        "description": "Retrieve problems for an ongoing contest.",
        "operationId": "contests-read_contest_problems",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestProblemsListPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contests/{id}": {
      "get": {
        "tags": [
          "contests"
        ],
        "summary": "Read Contest",
        "description": "Get Contest by ID.",
        "operationId": "contests-read_Contest",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "contests"
        ],
        "summary": "Update Contest",
        "description": "Update an Contest.",
        "operationId": "contests-update_contest",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contests"
        ],
        "summary": "Delete Contest",
        "description": "Delete an Contest.",
        "operationId": "contests-delete_contest",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/": {
      "get": {
        "tags": [
          "problems"
        ],
        "summary": "Read Problems",
        "description": "Retrieve Problems.",
        "operationId": "problems-read_problems",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "problems"
        ],
        "summary": "Create Problem",
        "description": "Create new Problem.",
        "operationId": "problems-create_problem",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{id}": {
      "get": {
        "tags": [
          "problems"
        ],
        "summary": "Read Problem",
        "description": "Get Problem by ID.",
        "operationId": "problems-read_problem",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "problems"
        ],
        "summary": "Update Problem",
        "description": "Update an Problem.",
        "operationId": "problems-update_problem",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "problems"
        ],
        "summary": "Delete Problem",
        "description": "Delete an Problem.",
        "operationId": "problems-delete_problem",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contest_problems/": {
      "post": {
        "tags": [
          "contest_problems"
        ],
        "summary": "Create Contest Problems",
        "description": "Create new ContestProblems.",
        "operationId": "contest_problems-create_contest_problems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestProblemsCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestProblemsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/contest_problems/{id}": {
      "put": {
        "tags": [
          "contest_problems"
        ],
        "summary": "Update Contest Problems",
        "description": "Update an ContestProblems.",
        "operationId": "contest_problems-update_contest_problems",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContestProblemsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContestProblemsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contest_problems"
        ],
        "summary": "Delete Contest Problems",
        "description": "Delete an ContestProblems.",
        "operationId": "contest_problems-delete_contest_problems",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContestCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "start_at",
          "end_at"
        ],
        "title": "ContestCreate"
      },
      "ContestProblemsCreate": {
        "properties": {
          "problem_id": {
            "type": "string",
            "format": "uuid",
            "title": "Problem Id"
          },
          "contest_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contest Id"
          },
          "order_num": {
            "type": "integer",
            "title": "Order Num",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "contest_id"
        ],
        "title": "ContestProblemsCreate"
      },
      "ContestProblemsListPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ContestProblemsPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "ContestProblemsListPublic"
      },
      "ContestProblemsPublic": {
        "properties": {
          "problem_id": {
            "type": "string",
            "format": "uuid",
            "title": "Problem Id"
          },
          "contest_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contest Id"
          },
          "order_num": {
            "type": "integer",
            "title": "Order Num",
            "default": 0
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "problem": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemMinimal"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "contest_id",
          "id"
        ],
        "title": "ContestProblemsPublic"
      },
      "ContestProblemsUpdate": {
        "properties": {
          "problem_id": {
            "type": "string",
            "format": "uuid",
            "title": "Problem Id"
          },
          "contest_id": {
            "type": "string",
            "format": "uuid",
            "title": "Contest Id"
          },
          "order_num": {
            "type": "integer",
            "title": "Order Num",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "problem_id",
          "contest_id"
        ],
        "title": "ContestProblemsUpdate"
      },
      "ContestPublic": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "problem_links": {
            "items": {
              "$ref": "#/components/schemas/ContestProblemsPublic"
            },
            "type": "array",
            "title": "Problem Links",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "start_at",
          "end_at",
          "id"
        ],
        "title": "ContestPublic"
      },
...
      "ContestUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "end_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "start_at",
          "end_at"
        ],
        "title": "ContestUpdate"
      },
      "ContestsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ContestPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "ContestsPublic"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "ProblemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "time_limit": {
            "type": "number",
            "maximum": 2000.0,
            "exclusiveMinimum": 0.0,
            "title": "Time Limit"
          },
          "memory_limit": {
            "type": "integer",
            "title": "Memory Limit"
          },
          "content": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content"
          },
          "input_format": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Input Format"
          },
          "output_format": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Output Format"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TestCaseSample"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 3,
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "name",
          "time_limit",
          "memory_limit",
          "content",
          "input_format",
          "output_format",
          "samples"
        ],
        "title": "ProblemCreate"
      },
...
      "ProblemPublic": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "time_limit": {
            "type": "number",
            "maximum": 2000.0,
            "exclusiveMinimum": 0.0,
            "title": "Time Limit"
          },
          "memory_limit": {
            "type": "integer",
            "title": "Memory Limit"
          },
          "content": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content"
          },
          "input_format": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Input Format"
          },
          "output_format": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Output Format"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/TestCaseSample"
            },
            "type": "array",
            "title": "Samples"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "time_limit",
          "memory_limit",
          "content",
          "input_format",
          "output_format",
          "id",
          "samples"
        ],
        "title": "ProblemPublic"
      },
      "ProblemUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "time_limit": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit"
          },
          "memory_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Limit"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "input_format": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Format"
          },
          "output_format": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Format"
          },
          "samples": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TestCaseSample"
                },
                "type": "array",
                "maxItems": 3,
                "minItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Samples"
          }
        },
        "type": "object",
        "title": "ProblemUpdate"
      },
      "ProblemsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProblemPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "ProblemsPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}